refactor: Hoist common functions
