+++ dotcom-rendering/src/web/browser/whenIdle.ts
