+++ dotcom-rendering/src/web/browser/whenVisible.ts
